+++ srcs/vector/ft_vunused_size.c
@@ -6,13 +6,13 @@
 /*   By: ahugh <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/26 15:01:37 by ahugh             #+#    #+#             */
-/*   Updated: 2019/02/26 15:04:10 by ahugh            ###   ########.fr       */
+/*   Updated: 2019/11/02 13:01:25 by ahugh            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t				ft_vunused_size(t_vector *v)
+long				ft_vunused_size(t_vector *v)
 {
 	if (v)
 		return (v->size - v->head * v->type_size);
